Move duplication item selectors into folder
