+++ client/app/bundles/course/duplication/pages/Duplication/ItemsSelector/AssessmentsSelector.jsx
