+++ client/app/bundles/course/duplication/pages/Duplication/ItemsSelector/MaterialsSelector.jsx
