+++ client/app/bundles/course/duplication/pages/Duplication/ItemsSelector/SurveysSelector.jsx
