+++ client/app/bundles/course/duplication/pages/Duplication/ItemsSelector/VideosSelector.jsx
